requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2==2.9.10
python-dotenv==1.0.1
drf-yasg==1.21.8
